backend/pyproject.toml
[project]
name = "portfolio-backend"
version = "0.1.0"
description = "Portfolio analysis backend (FastAPI + yfinance)"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.136.0",
    "numpy>=2.2.6",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.2",
    "scipy>=1.15.3",
    "uvicorn>=0.44.0",
    "yfinance>=1.3.0",
]

[tool.uv]
package = false
